♻️ Refactor with Header, Hero and Footer components
