+++ app.vue
@@ -1,544 +1,20 @@
 <template>
-  <div class="bg-white">
+  <div>
     <!-- Header -->
-    <header class="absolute inset-x-0 top-0 z-50">
-      <nav class="flex items-center justify-between p-6 lg:px-8" aria-label="Global">
-        <div class="flex lg:flex-1">
-          <a href="#" class="-m-1.5 p-1.5">
-            <span class="sr-only">anthonypillotOS</span>
-            <NuxtImg
-              format="webp"
-              quality="80"
-              class="h-12 w-auto"
-              src="https://raw.githubusercontent.com/anthonypillot/assets/78c3327a009fa95b72dd06dac4bbc3e811f5a1c3/logo/svg/logo_anthonypillotOS_OS.svg"
-              alt="anthonypillotOS"
-              style="filter: invert(1)"
-            />
-          </a>
-        </div>
-        <div class="flex lg:hidden">
-          <button
-            type="button"
-            class="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-400"
-            @click="mobileMenuOpen = true"
-          >
-            <span class="sr-only">Open main menu</span>
-            <Bars3Icon class="h-6 w-6" aria-hidden="true" />
-          </button>
-        </div>
-        <div class="hidden lg:flex lg:gap-x-12">
-          <a
-            v-for="item in navigation"
-            :key="item.name"
-            :href="item.href"
-            :rel="item.rel"
-            :target="item.target"
-            class="text-sm font-semibold leading-6 text-white"
-            >{{ item.name }}</a
-          >
-        </div>
-        <div class="hidden lg:flex lg:flex-1 lg:justify-end">
-          <NuxtLink :href="data.url.githubRepository" rel="noopener" target="_blank" class="text-sm font-semibold leading-6 text-white"
-            >GitHub repository <span aria-hidden="true">&rarr;</span></NuxtLink
-          >
-        </div>
-      </nav>
-      <Dialog as="div" class="lg:hidden" @close="mobileMenuOpen = false" :open="mobileMenuOpen">
-        <div class="fixed inset-0 z-50" />
-        <DialogPanel
-          class="fixed inset-y-0 right-0 z-50 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10"
-        >
-          <div class="flex items-center justify-between">
-            <a href="#" class="-m-1.5 p-1.5">
-              <span class="sr-only">anthonypillotOS</span>
-              <NuxtImg
-                format="webp"
-                quality="80"
-                class="h-12 w-auto"
-                src="https://raw.githubusercontent.com/anthonypillot/assets/78c3327a009fa95b72dd06dac4bbc3e811f5a1c3/logo/svg/logo_anthonypillotOS_OS.svg"
-                alt=""
-              />
-            </a>
-            <button type="button" class="-m-2.5 rounded-md p-2.5 text-gray-700" @click="mobileMenuOpen = false">
-              <span class="sr-only">Close menu</span>
-              <XMarkIcon class="h-6 w-6" aria-hidden="true" />
-            </button>
-          </div>
-          <div class="mt-6 flow-root">
-            <div class="-my-6 divide-y divide-gray-500/10">
-              <div class="space-y-2 py-6">
-                <a
-                  v-for="item in navigation"
-                  :key="item.name"
-                  :href="item.href"
-                  :rel="item.rel"
-                  :target="item.target"
-                  class="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
-                  >{{ item.name }}</a
-                >
-              </div>
-              <div class="py-6">
-                <a
-                  :href="data.url.githubRepository"
-                  rel="noopener"
-                  target="_blank"
-                  class="-mx-3 block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
-                  >GitHub repository &rarr;</a
-                >
-              </div>
-            </div>
-          </div>
-        </DialogPanel>
-      </Dialog>
-    </header>
+    <BaseHeader :data="data" />
 
+    <!-- Main content -->
     <main>
       <!-- Hero section -->
-      <div class="relative isolate overflow-hidden bg-gray-900 pb-16 pt-14 sm:pb-20">
-        <NuxtImg
-          format="webp"
-          quality="80"
-          src="https://focus.courrierinternational.com/2022/02/04/0/0/3882/2584/1280/0/60/0/89816dd_1644014034247-christopher-gower-m-hrflhgabo-unsplash.jpg"
-          alt=""
-          class="absolute inset-0 -z-10 h-full w-full object-cover"
-          style="filter: blur(4px) hue-rotate(60deg) brightness(25%)"
-        />
-        <div class="absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80" aria-hidden="true">
-          <div
-            class="relative left-[calc(50%-11rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-20 sm:left-[calc(50%-30rem)] sm:w-[72.1875rem]"
-            style="
-              clip-path: polygon(
-                74.1% 44.1%,
-                100% 61.6%,
-                97.5% 26.9%,
-                85.5% 0.1%,
-                80.7% 2%,
-                72.5% 32.5%,
-                60.2% 62.4%,
-                52.4% 68.1%,
-                47.5% 58.3%,
-                45.2% 34.5%,
-                27.5% 76.7%,
-                0.1% 64.9%,
-                17.9% 100%,
-                27.6% 76.8%,
-                76.1% 97.7%,
-                74.1% 44.1%
-              );
-            "
-          />
-        </div>
-        <div class="mx-auto max-w-7xl px-6 lg:px-8">
-          <div class="mx-auto max-w-2xl py-32 sm:py-48 lg:py-32">
-            <div class="hidden sm:mb-8 sm:flex sm:justify-center">
-              <div class="relative rounded-full px-3 py-1 text-sm leading-6 text-gray-400 ring-1 ring-white/10 hover:ring-white/20">
-                Discover our Size Up organization documentation website.
-                <NuxtLink :href="data.url.sizeUpDocumentation" rel="noopener" target="_blank" class="font-semibold text-white"
-                  ><span class="absolute inset-0" aria-hidden="true" />Get me in <span aria-hidden="true">&rarr;</span></NuxtLink
-                >
-              </div>
-            </div>
-            <div class="flex flex-col text-center">
-              <NuxtImg
-                format="webp"
-                quality="80"
-                class="h-80"
-                src="https://raw.githubusercontent.com/anthonypillot/assets/main/logo/svg/logo_anthonypillotOS_white.svg"
-                alt="anthonypillotOS"
-              />
-              <h1 class="hidden text-4xl font-bold tracking-tight text-white sm:text-6xl">anthonypillotOS</h1>
-              <p class="mt-6 text-lg leading-8 text-gray-300">
-                <span class="text-indigo-400">Freelance</span> Software Engineer, <span class="text-indigo-400">Real</span> Full Stack
-                Developer, and <span class="text-indigo-400">more</span>.
-              </p>
-              <p class="mt-6 text-lg leading-8 text-gray-300">
-                <span class="text-indigo-400">Lille</span>, France. <span class="text-indigo-400">Remote</span> and
-                <span class="text-indigo-400">on-site</span> work.
-              </p>
-              <!-- <div class="mt-10 flex items-center justify-center gap-x-6">
-                <a
-                  href="#"
-                  class="rounded-md bg-indigo-500 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-indigo-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-400"
-                  >Get started</a
-                >
-                <a href="#" class="text-sm font-semibold leading-6 text-white">Live demo <span aria-hidden="true">→</span></a>
-              </div> -->
-            </div>
-          </div>
-
-          <!-- Logo cloud -->
-          <div
-            class="mx-auto grid max-w-lg grid-cols-4 items-center gap-x-8 gap-y-10 sm:max-w-xl sm:grid-cols-6 sm:gap-x-10 lg:mx-0 lg:max-w-none lg:grid-cols-5"
-          >
-            <NuxtImg
-              format="webp"
-              quality="80"
-              class="col-span-2 max-h-12 w-full object-contain sm:col-start-2 lg:col-span-1"
-              src="https://upload.wikimedia.org/wikipedia/commons/4/4c/Typescript_logo_2020.svg"
-              alt="TypeScript"
-              width="158"
-              height="48"
-              style="filter: hue-rotate(30deg)"
-            />
-            <NuxtImg
-              format="webp"
-              quality="80"
-              class="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
-              src="https://upload.wikimedia.org/wikipedia/commons/9/95/Vue.js_Logo_2.svg"
-              alt="Vue.js"
-              width="158"
-              height="48"
-              style="filter: hue-rotate(110deg)"
-            />
-            <NuxtImg
-              format="webp"
-              quality="80"
-              class="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
-              src="https://upload.wikimedia.org/wikipedia/commons/a/ae/Nuxt_logo.svg"
-              alt="Nuxt"
-              width="158"
-              height="48"
-              style="filter: hue-rotate(110deg)"
-            />
-            <NuxtImg
-              format="webp"
-              quality="80"
-              class="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
-              src="https://static-00.iconduck.com/assets.00/node-js-icon-454x512-nztofx17.png"
-              alt="Node.js"
-              width="158"
-              height="48"
-              style="filter: hue-rotate(145deg)"
-            />
-            <NuxtImg
-              format="webp"
-              quality="80"
-              class="col-span-2 col-start-2 max-h-12 w-full object-contain sm:col-start-auto lg:col-span-1"
-              src="https://www.vectorlogo.zone/logos/java/java-icon.svg"
-              alt="Java"
-              width="158"
-              height="48"
-              style="filter: hue-rotate(220deg)"
-            />
-          </div>
-        </div>
-        <div
-          class="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]"
-          aria-hidden="true"
-        >
-          <div
-            class="relative left-[calc(50%+3rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-20 sm:left-[calc(50%+36rem)] sm:w-[72.1875rem]"
-            style="
-              clip-path: polygon(
-                74.1% 44.1%,
-                100% 61.6%,
-                97.5% 26.9%,
-                85.5% 0.1%,
-                80.7% 2%,
-                72.5% 32.5%,
-                60.2% 62.4%,
-                52.4% 68.1%,
-                47.5% 58.3%,
-                45.2% 34.5%,
-                27.5% 76.7%,
-                0.1% 64.9%,
-                17.9% 100%,
-                27.6% 76.8%,
-                76.1% 97.7%,
-                74.1% 44.1%
-              );
-            "
-          />
-        </div>
-      </div>
-
-      <!-- Feature section -->
-      <!-- <div class="mt-32 sm:mt-56">
-        <div class="mx-auto max-w-7xl px-6 lg:px-8">
-          <div class="mx-auto max-w-2xl sm:text-center">
-            <h2 class="text-base font-semibold leading-7 text-indigo-600">Everything you need</h2>
-            <p class="mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">No server? No problem.</p>
-            <p class="mt-6 text-lg leading-8 text-gray-600">
-              Lorem ipsum, dolor sit amet consectetur adipisicing elit. Maiores impedit perferendis suscipit eaque, iste dolor cupiditate
-              blanditiis.
-            </p>
-          </div>
-        </div>
-        <div class="relative overflow-hidden pt-16">
-          <div class="mx-auto max-w-7xl px-6 lg:px-8">
-            <NuxtImg format="webp"
-quality="80"
-              src="https://tailwindui.com/img/component-images/project-app-screenshot.png"
-              alt="App screenshot"
-              class="mb-[-12%] rounded-xl shadow-2xl ring-1 ring-gray-900/10"
-              width="2432"
-              height="1442"
-            />
-            <div class="relative" aria-hidden="true">
-              <div class="absolute -inset-x-20 bottom-0 bg-gradient-to-t from-white pt-[7%]" />
-            </div>
-          </div>
-        </div>
-        <div class="mx-auto mt-16 max-w-7xl px-6 sm:mt-20 md:mt-24 lg:px-8">
-          <dl
-            class="mx-auto grid max-w-2xl grid-cols-1 gap-x-6 gap-y-10 text-base leading-7 text-gray-600 sm:grid-cols-2 lg:mx-0 lg:max-w-none lg:grid-cols-3 lg:gap-x-8 lg:gap-y-16"
-          >
-            <div v-for="feature in features" :key="feature.name" class="relative pl-9">
-              <dt class="inline font-semibold text-gray-900">
-                <component :is="feature.icon" class="absolute left-1 top-1 h-5 w-5 text-indigo-600" aria-hidden="true" />
-                {{ feature.name }}
-              </dt>
-              {{ " " }}
-              <dd class="inline">{{ feature.description }}</dd>
-            </div>
-          </dl>
-        </div>
-      </div> -->
-
-      <!-- Testimonial section -->
-      <!-- <div class="relative z-10 mt-32 bg-gray-900 pb-20 sm:mt-56 sm:pb-24 xl:pb-0">
-        <div class="absolute inset-0 overflow-hidden" aria-hidden="true">
-          <div class="absolute left-[calc(50%-19rem)] top-[calc(50%-36rem)] transform-gpu blur-3xl">
-            <div
-              class="aspect-[1097/1023] w-[68.5625rem] bg-gradient-to-r from-[#ff4694] to-[#776fff] opacity-25"
-              style="
-                clip-path: polygon(
-                  74.1% 44.1%,
-                  100% 61.6%,
-                  97.5% 26.9%,
-                  85.5% 0.1%,
-                  80.7% 2%,
-                  72.5% 32.5%,
-                  60.2% 62.4%,
-                  52.4% 68.1%,
-                  47.5% 58.3%,
-                  45.2% 34.5%,
-                  27.5% 76.7%,
-                  0.1% 64.9%,
-                  17.9% 100%,
-                  27.6% 76.8%,
-                  76.1% 97.7%,
-                  74.1% 44.1%
-                );
-              "
-            />
-          </div>
-        </div>
-        <div class="mx-auto flex max-w-7xl flex-col items-center gap-x-8 gap-y-10 px-6 sm:gap-y-8 lg:px-8 xl:flex-row xl:items-stretch">
-          <div class="-mt-8 w-full max-w-2xl xl:-mb-8 xl:w-96 xl:flex-none">
-            <div class="relative aspect-[2/1] h-full md:-mx-8 xl:mx-0 xl:aspect-auto">
-              <NuxtImg format="webp"
-quality="80"
-                class="absolute inset-0 h-full w-full rounded-2xl bg-gray-800 object-cover shadow-2xl"
-                src="https://images.unsplash.com/photo-1517841905240-472988babdf9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2102&q=80"
-                alt=""
-              />
-            </div>
-          </div>
-          <div class="w-full max-w-2xl xl:max-w-none xl:flex-auto xl:px-16 xl:py-24">
-            <figure class="relative isolate pt-6 sm:pt-12">
-              <svg viewBox="0 0 162 128" fill="none" aria-hidden="true" class="absolute left-0 top-0 -z-10 h-32 stroke-white/20">
-                <path
-                  id="b56e9dab-6ccb-4d32-ad02-6b4bb5d9bbeb"
-                  d="M65.5697 118.507L65.8918 118.89C68.9503 116.314 71.367 113.253 73.1386 109.71C74.9162 106.155 75.8027 102.28 75.8027 98.0919C75.8027 94.237 75.16 90.6155 73.8708 87.2314C72.5851 83.8565 70.8137 80.9533 68.553 78.5292C66.4529 76.1079 63.9476 74.2482 61.0407 72.9536C58.2795 71.4949 55.276 70.767 52.0386 70.767C48.9935 70.767 46.4686 71.1668 44.4872 71.9924L44.4799 71.9955L44.4726 71.9988C42.7101 72.7999 41.1035 73.6831 39.6544 74.6492C38.2407 75.5916 36.8279 76.455 35.4159 77.2394L35.4047 77.2457L35.3938 77.2525C34.2318 77.9787 32.6713 78.3634 30.6736 78.3634C29.0405 78.3634 27.5131 77.2868 26.1274 74.8257C24.7483 72.2185 24.0519 69.2166 24.0519 65.8071C24.0519 60.0311 25.3782 54.4081 28.0373 48.9335C30.703 43.4454 34.3114 38.345 38.8667 33.6325C43.5812 28.761 49.0045 24.5159 55.1389 20.8979C60.1667 18.0071 65.4966 15.6179 71.1291 13.7305C73.8626 12.8145 75.8027 10.2968 75.8027 7.38572C75.8027 3.6497 72.6341 0.62247 68.8814 1.1527C61.1635 2.2432 53.7398 4.41426 46.6119 7.66522C37.5369 11.6459 29.5729 17.0612 22.7236 23.9105C16.0322 30.6019 10.618 38.4859 6.47981 47.558L6.47976 47.558L6.47682 47.5647C2.4901 56.6544 0.5 66.6148 0.5 77.4391C0.5 84.2996 1.61702 90.7679 3.85425 96.8404L3.8558 96.8445C6.08991 102.749 9.12394 108.02 12.959 112.654L12.959 112.654L12.9646 112.661C16.8027 117.138 21.2829 120.739 26.4034 123.459L26.4033 123.459L26.4144 123.465C31.5505 126.033 37.0873 127.316 43.0178 127.316C47.5035 127.316 51.6783 126.595 55.5376 125.148L55.5376 125.148L55.5477 125.144C59.5516 123.542 63.0052 121.456 65.9019 118.881L65.5697 118.507Z"
-                />
-                <use href="#b56e9dab-6ccb-4d32-ad02-6b4bb5d9bbeb" x="86" />
-              </svg>
-              <blockquote class="text-xl font-semibold leading-8 text-white sm:text-2xl sm:leading-9">
-                <p>
-                  Gravida quam mi erat tortor neque molestie. Auctor aliquet at porttitor a enim nunc suscipit tincidunt nunc. Et non lorem
-                  tortor posuere. Nunc eu scelerisque interdum eget tellus non nibh scelerisque bibendum.
-                </p>
-              </blockquote>
-              <figcaption class="mt-8 text-base">
-                <div class="font-semibold text-white">Judith Black</div>
-                <div class="mt-1 text-gray-400">CEO of Tuple</div>
-              </figcaption>
-            </figure>
-          </div>
-        </div>
-      </div> -->
-
-      <!-- Pricing section -->
-      <!-- <div class="relative isolate mt-32 bg-white px-6 sm:mt-56 lg:px-8">
-        <div class="absolute inset-x-0 -top-3 -z-10 transform-gpu overflow-hidden px-36 blur-3xl" aria-hidden="true">
-          <div
-            class="mx-auto aspect-[1155/678] w-[72.1875rem] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30"
-            style="
-              clip-path: polygon(
-                74.1% 44.1%,
-                100% 61.6%,
-                97.5% 26.9%,
-                85.5% 0.1%,
-                80.7% 2%,
-                72.5% 32.5%,
-                60.2% 62.4%,
-                52.4% 68.1%,
-                47.5% 58.3%,
-                45.2% 34.5%,
-                27.5% 76.7%,
-                0.1% 64.9%,
-                17.9% 100%,
-                27.6% 76.8%,
-                76.1% 97.7%,
-                74.1% 44.1%
-              );
-            "
-          />
-        </div>
-        <div class="mx-auto max-w-2xl text-center lg:max-w-4xl">
-          <h2 class="text-base font-semibold leading-7 text-indigo-600">Pricing</h2>
-          <p class="mt-2 text-4xl font-bold tracking-tight text-gray-900 sm:text-5xl">The right price for you, whoever you are</p>
-        </div>
-        <p class="mx-auto mt-6 max-w-2xl text-center text-lg leading-8 text-gray-600">
-          Qui iusto aut est earum eos quae. Eligendi est at nam aliquid ad quo reprehenderit in aliquid fugiat dolorum voluptatibus.
-        </p>
-        <div class="mx-auto mt-16 grid max-w-lg grid-cols-1 items-center gap-y-6 sm:mt-20 sm:gap-y-0 lg:max-w-4xl lg:grid-cols-2">
-          <div
-            v-for="(tier, tierIdx) in tiers"
-            :key="tier.id"
-            :class="[
-              tier.featured ? 'relative bg-gray-900 shadow-2xl' : 'bg-white/60 sm:mx-8 lg:mx-0',
-              tier.featured
-                ? ''
-                : tierIdx === 0
-                ? 'rounded-t-3xl sm:rounded-b-none lg:rounded-tr-none lg:rounded-bl-3xl'
-                : 'sm:rounded-t-none lg:rounded-tr-3xl lg:rounded-bl-none',
-              'rounded-3xl p-8 ring-1 ring-gray-900/10 sm:p-10',
-            ]"
-          >
-            <h3 :id="tier.id" :class="[tier.featured ? 'text-indigo-400' : 'text-indigo-600', 'text-base font-semibold leading-7']">
-              {{ tier.name }}
-            </h3>
-            <p class="mt-4 flex items-baseline gap-x-2">
-              <span :class="[tier.featured ? 'text-white' : 'text-gray-900', 'text-5xl font-bold tracking-tight']">{{
-                tier.priceMonthly
-              }}</span>
-              <span :class="[tier.featured ? 'text-gray-400' : 'text-gray-500', 'text-base']">/month</span>
-            </p>
-            <p :class="[tier.featured ? 'text-gray-300' : 'text-gray-600', 'mt-6 text-base leading-7']">{{ tier.description }}</p>
-            <ul role="list" :class="[tier.featured ? 'text-gray-300' : 'text-gray-600', 'mt-8 space-y-3 text-sm leading-6 sm:mt-10']">
-              <li v-for="feature in tier.features" :key="feature" class="flex gap-x-3">
-                <CheckIcon :class="[tier.featured ? 'text-indigo-400' : 'text-indigo-600', 'h-6 w-5 flex-none']" aria-hidden="true" />
-                {{ feature }}
-              </li>
-            </ul>
-            <a
-              :href="tier.href"
-              :aria-describedby="tier.id"
-              :class="[
-                tier.featured
-                  ? 'bg-indigo-500 text-white shadow-sm hover:bg-indigo-400 focus-visible:outline-indigo-500'
-                  : 'text-indigo-600 ring-1 ring-inset ring-indigo-200 hover:ring-indigo-300 focus-visible:outline-indigo-600',
-                'mt-8 block rounded-md py-2.5 px-3.5 text-center text-sm font-semibold focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 sm:mt-10',
-              ]"
-              >Get started today</a
-            >
-          </div>
-        </div>
-      </div> -->
-
-      <!-- FAQ section -->
-      <!-- <div class="mx-auto mt-32 max-w-7xl px-6 sm:mt-56 lg:px-8">
-        <div class="mx-auto max-w-4xl divide-y divide-gray-900/10">
-          <h2 class="text-2xl font-bold leading-10 tracking-tight text-gray-900">Frequently asked questions</h2>
-          <dl class="mt-10 space-y-6 divide-y divide-gray-900/10">
-            <Disclosure as="div" v-for="faq in faqs" :key="faq.question" class="pt-6" v-slot="{ open }">
-              <dt>
-                <DisclosureButton class="flex w-full items-start justify-between text-left text-gray-900">
-                  <span class="text-base font-semibold leading-7">{{ faq.question }}</span>
-                  <span class="ml-6 flex h-7 items-center">
-                    <PlusSmallIcon v-if="!open" class="h-6 w-6" aria-hidden="true" />
-                    <MinusSmallIcon v-else class="h-6 w-6" aria-hidden="true" />
-                  </span>
-                </DisclosureButton>
-              </dt>
-              <DisclosurePanel as="dd" class="mt-2 pr-12">
-                <p class="text-base leading-7 text-gray-600">{{ faq.answer }}</p>
-              </DisclosurePanel>
-            </Disclosure>
-          </dl>
-        </div>
-      </div> -->
+      <BaseHero :data="data" />
     </main>
 
     <!-- Footer -->
-    <footer class="bg-gray-900" aria-labelledby="footer-heading">
-      <h2 id="footer-heading" class="sr-only">Footer</h2>
-      <div class="mx-auto max-w-7xl px-6 py-16 sm:py-24 lg:px-8 lg:py-32">
-        <div class="xl:grid xl:grid-cols-3 xl:gap-8">
-          <NuxtImg
-            format="webp"
-            quality="80"
-            class="h-7"
-            src="https://raw.githubusercontent.com/anthonypillot/assets/78c3327a009fa95b72dd06dac4bbc3e811f5a1c3/logo/svg/logo_anthonypillotOS_OS.svg"
-            alt="anthonypillotOS"
-            style="filter: invert(1)"
-          />
-          <div class="mt-16 grid grid-cols-2 gap-8 xl:col-span-2 xl:mt-0">
-            <div class="md:grid md:grid-cols-2 md:gap-8">
-              <div>
-                <h3 class="text-sm font-semibold leading-6 text-white">GitHub</h3>
-                <ul role="list" class="mt-6 space-y-4">
-                  <li v-for="item in footerNavigation.github" :key="item.name">
-                    <a :href="item.href" :rel="item.rel" :target="item.target" class="text-sm leading-6 text-gray-300 hover:text-white">{{
-                      item.name
-                    }}</a>
-                  </li>
-                </ul>
-              </div>
-              <div class="mt-10 md:mt-0">
-                <h3 class="text-sm font-semibold leading-6 text-white">Project</h3>
-                <ul role="list" class="mt-6 space-y-4">
-                  <li v-for="item in footerNavigation.project" :key="item.name">
-                    <a :href="item.href" :rel="item.rel" :target="item.target" class="text-sm leading-6 text-gray-300 hover:text-white">{{
-                      item.name
-                    }}</a>
-                  </li>
-                </ul>
-              </div>
-            </div>
-            <div class="md:grid md:grid-cols-2 md:gap-8">
-              <div>
-                <h3 class="text-sm font-semibold leading-6 text-white">Contact</h3>
-                <ul role="list" class="mt-6 space-y-4">
-                  <li v-for="item in footerNavigation.contact" :key="item.name">
-                    <a :href="item.href" :rel="item.rel" :target="item.target" class="text-sm leading-6 text-gray-300 hover:text-white">{{
-                      item.name
-                    }}</a>
-                  </li>
-                </ul>
-              </div>
-              <!-- <div class="mt-10 md:mt-0">
-                <h3 class="text-sm font-semibold leading-6 text-white">Legal</h3>
-                <ul role="list" class="mt-6 space-y-4">
-                  <li v-for="item in footerNavigation.legal" :key="item.name">
-                    <a :href="item.href" class="text-sm leading-6 text-gray-300 hover:text-white">{{ item.name }}</a>
-                  </li>
-                </ul>
-              </div> -->
-            </div>
-          </div>
-        </div>
-      </div>
-    </footer>
+    <BaseFooter :data="data" />
   </div>
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
-import { Dialog, DialogPanel, Disclosure, DisclosureButton, DisclosurePanel } from "@headlessui/vue";
-import { Bars3Icon, MinusSmallIcon, PlusSmallIcon, XMarkIcon } from "@heroicons/vue/24/outline";
-import {
-  ArrowPathIcon,
-  CheckIcon,
-  CloudArrowUpIcon,
-  Cog6ToothIcon,
-  FingerPrintIcon,
-  LockClosedIcon,
-  ServerIcon,
-} from "@heroicons/vue/20/solid";
-
 const data = {
   email: {
     owner: "pillot.anthony@gmail.com",
@@ -552,104 +28,4 @@ const data = {
     sizeUpDocumentation: "https://docs.sizeup.cloud",
   },
 };
-
-const navigation = [
-  { name: "Home", href: "/", rel: "", target: "" },
-  { name: "My LinkedIn", href: data.url.linkedIn, rel: "noopener", target: "_blank" },
-  { name: "My GitHub", href: data.url.githubAccount, rel: "noopener", target: "_blank" },
-  { name: "Size Up - Org.", href: data.url.githubOrganization, rel: "noopener", target: "_blank" },
-  { name: "Size Up - Documentation", href: data.url.sizeUpDocumentation, rel: "noopener", target: "_blank" },
-];
-
-// const features = [
-//   {
-//     name: "Push to deploy.",
-//     description: "Lorem ipsum, dolor sit amet consectetur adipisicing elit aute id magna.",
-//     icon: CloudArrowUpIcon,
-//   },
-//   {
-//     name: "SSL certificates.",
-//     description: "Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui lorem cupidatat commodo.",
-//     icon: LockClosedIcon,
-//   },
-//   {
-//     name: "Simple queues.",
-//     description: "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus.",
-//     icon: ArrowPathIcon,
-//   },
-//   {
-//     name: "Advanced security.",
-//     description: "Lorem ipsum, dolor sit amet consectetur adipisicing elit aute id magna.",
-//     icon: FingerPrintIcon,
-//   },
-//   {
-//     name: "Powerful API.",
-//     description: "Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui lorem cupidatat commodo.",
-//     icon: Cog6ToothIcon,
-//   },
-//   {
-//     name: "Database backups.",
-//     description: "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. ",
-//     icon: ServerIcon,
-//   },
-// ];
-// const tiers = [
-//   {
-//     name: "Hobby",
-//     id: "tier-hobby",
-//     href: "#",
-//     priceMonthly: "$19",
-//     description: "The perfect plan if you're just getting started with our product.",
-//     features: ["25 products", "Up to 10,000 subscribers", "Advanced analytics", "24-hour support response time"],
-//     featured: false,
-//   },
-//   {
-//     name: "Enterprise",
-//     id: "tier-enterprise",
-//     href: "#",
-//     priceMonthly: "$49",
-//     description: "Dedicated support and infrastructure for your company.",
-//     features: [
-//       "Unlimited products",
-//       "Unlimited subscribers",
-//       "Advanced analytics",
-//       "Dedicated support representative",
-//       "Marketing automations",
-//       "Custom integrations",
-//     ],
-//     featured: true,
-//   },
-// ];
-// const faqs = [
-//   {
-//     question: "What's the best thing about Switzerland?",
-//     answer:
-//       "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
-//   },
-//   // More questions...
-// ];
-
-const footerNavigation = {
-  github: [
-    { name: "Account", href: data.url.githubAccount, rel: "noopener", target: "_blank" },
-    { name: "Repository", href: data.url.githubRepository, rel: "noopener", target: "_blank" },
-    { name: "Organization", href: data.url.githubOrganization, rel: "noopener", target: "_blank" },
-  ],
-  project: [
-    { name: "Size Up - Documentation", href: data.url.sizeUpDocumentation, rel: "noopener", target: "_blank" },
-    { name: "Free Games Catcher", href: data.url.githubFreeGamesCatcherCore, rel: "noopener", target: "_blank" },
-  ],
-  contact: [
-    { name: "Email", href: "mailto:" + data.email.owner },
-    { name: "LinkedIn", href: data.url.linkedIn, rel: "noopener", target: "_blank" },
-    { name: "GitHub", href: data.url.githubAccount, rel: "noopener", target: "_blank" },
-  ],
-  // legal: [
-  //   { name: "Claim", href: "#" },
-  //   { name: "Privacy", href: "#" },
-  //   { name: "Terms", href: "#" },
-  // ],
-};
-
-const mobileMenuOpen: Ref<boolean> = ref(false);
 </script>
